회원가입을 위한 코드 추가

+++ src/main/java/com/muglove/projcet/muglove/entity/JoinRepository.java
@@ -4,6 +4,6 @@ package com.muglove.projcet.muglove.entity;
 import com.muglove.projcet.muglove.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JoinRepository extends JpaRepository<Member, Long> {
+public interface JoinRepository extends JpaRepository<Member, Integer> {
 
 }
